+++ lib/flutter_metrics_sdk.dart
@@ -10,6 +10,7 @@ library;
 export 'src/metrics_client.dart';
 export 'src/metric_record.dart';
 export 'src/metrics_event.dart';
+export 'src/screen_context.dart';
 export 'src/screen_tracker.dart';
 export 'src/crash_tracker.dart';
 export 'src/api_tracker.dart';